Consolidated module list functionality into standalone component
Enact-DCO-1.0-Signed-off-by: Blake Stephens <blake.stephens@lge.com>
